Migrate Compositor to use pmndrs/postprocessing for better performance
GPU time doesn't show any difference, but CPU time gets slight improvement
(~2ms), and the graph is more stable, especially on Windows.

For bloom, the BloomEffect from pmndrs/postprocessing is also based on
Unreal Engine's bloom effect (says docs), params are different so tried
to mimic the look previously done with UnrealBloom.

NOTE:
The bloom has a slight tint of orange when the effect is the strongest,
e.g, when the light is near ground (low Y) with grazing viewing angle.
It's unintended and don't know why it's happening, but it's closer to
the look in real life, so I'll leave it this way.