fix(DropHelper): Make bound checking / setting state in correct order
The state was incorrectly set to "active" (dragging = true,
UIControlState to "interacting") first before bound checking.

It was corrected by subsequent handler for `pointerup` event in the same
function (dragging = false), and `App` setting UIControlState to 'idle'
for `mouseup`.

This changes makes the state setting logic correctly by checking bound
first.

BTW, the event mismatch puts in a wrong state for pointer devices, which
will be fixed in a later commit.